메뉴 > 타이틀 버그 수정
메뉴 > 타이틀 갈시 플레이어 오브젝트가 꼬이는 점이 있어 수정
가상키보드의 엔터가 미작동하던점 수정
