--- Assets/Scripts/Mode/ClearTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a8a4fdf5332779841bb57d902c208bd1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
