+++ Assets/Scripts/Title/TitleManager.cs.meta
@@ -4,7 +4,7 @@ MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 150
+  executionOrder: 20
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
